// src/app/pokemon/interfaces/pkTypeInterface.ts
// Generated by https://quicktype.io

export interface IType {
    damage_relations:      DamageRelations;
    game_indices:          any[];
    generation:            Generation;
    id:                    number;
    move_damage_class:     Generation;
    moves:                 any[];
    name:                  string;
    names:                 Name[];
    past_damage_relations: any[];
    pokemon:               Pokemon[];
}

export interface DamageRelations {
    double_damage_from: DoubleDamageFrom[];
    double_damage_to:   DoubleDamageFrom[];
    half_damage_from:   DoubleDamageFrom[];
    half_damage_to:     DoubleDamageFrom[];
    no_damage_from:     DoubleDamageFrom[];
    no_damage_to:       any[];
}

export interface DoubleDamageFrom {
    name: string;
    url:  string;
}

export interface Generation {
}

export interface Name {
    language: DoubleDamageFrom;
    name:     string;
}

export interface Pokemon {
    pokemon: DoubleDamageFrom;
    slot:    number;
}